src/backend-api/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
pydantic-settings>=2.6,<3.0
alembic>=1.14,<2.0
PyJWT>=2.9,<3.0
